Add working day08 (this was a LOOOG one, I've lost hope at some point..)
For this one I went is a hard and un-necessarily complex solution first,
and it didn't work. I wanted to deduce which wire was driving which
segment, kind of like a Sudoku puzzle. THIS IS HARD..
I lost hope and didn't know how to do differently..

Finally I 'cheated' and checked how others did it, and they didn't even
tried to do that, but instead went on to match a signal pattern to a
number directly, and deduction based on the number of segments and
comparisons with known or common ones.

Finally I removed my initial (broken) impl, and re-written an impl impl
by that logic. And yes, it works \o/